+++ packages/eslint-config-base/CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.0.10 - 2021-06-08
+- Fix conflicting configuration with no-use-before-define and @typescript-eslint/no-use-before-define
+- Fix conflicting configuration with no-shadow and @typescript-estlin/no-shadow
+
+## 0.0.9 - 2021-06-07
+- Update eslint packages
+
 ## 0.0.5 - 2019-09-19
 - Update eslint packages